==> 0_analyses/0_exhaustive/3_rbind.ipynb <==
wd <- dirname(dirname(getwd()))
source(paste0(wd,"/mission_control/treasure_map.R"))
source(paste0(wd,"/mission_control/helpers/exhaustive_study/exhaustive_help.R"))
library(tidyverse)

### 0 - Gather results

results <- data.frame()
for( i in list.files(paste0(TMP_DIR, 'exhaustive_study/'))){
    results <- rbind(results, readRDS( paste0(TMP_DIR, 'exhaustive_study/',i)))
}

### 1 - Send it!

saveRDS( results, paste0( TMP_DIR, paste0( "exhaustive-stats.Rds" ) ))
