==> 0_analyses/0_exhaustive/6_plot-ingredients.ipynb <==
wd <- dirname(dirname(getwd()))
source(paste0(wd,"/mission_control/treasure_map.R"))
source(paste0(wd,"/mission_control/helpers/figures/exhaustive_prep.R"))
library(tidyverse)
library(stringr)
library(data.table)

### 0 - Read combine data

boom <- readRDS(paste0(TMP_DIR,"exhaustive-combine.Rds"))

### 1 - Possibly more filters

##### Filter 

boom <- boom %>% filter( log10_p > .01, col_type != "factor", feature != "pretreat_comp")
boom$est <- ifelse( boom$col_type == "factor", 0, boom$est)
boom$cor_pretreat <- ifelse( boom$feature == "clinical_pre_to_post_treatment_time", 0,  boom$cor_pretreat)

##### Add feature groups

boom$feature_group <- unlist(lapply(boom$feature, function(i) strsplit(i, "_")[[1]][1]))
boom$feature_group <- ifelse( boom$feature %in% c("tcell", "tgfb", "prolif", "pdl1"), "isofox", boom$feature_group)
boom$feature_group <- ifelse( boom$feature %in% c("sv", "tmb","purity"), "somatic", boom$feature_group)   
boom$feature_group <- ifelse( boom$feature %in% c("pretreat", "age"), "clinical", boom$feature_group)                                       

#### Set multiple test threshold

boom$p_adj_by <- p.adjust(boom$p_val, method = "BY")
boom$p_adj_bh <- p.adjust(boom$p_val, method = "BH")
boom$by_05_fdr <- boom %>% filter( p_adj_by < .05 ) %>% arrange( log10_p ) %>% head(1) %>% pull(p_val)
boom$bh_05_fdr <- boom %>% filter( p_adj_bh < .05 ) %>% arrange( log10_p ) %>% head(1) %>% pull(p_val)
boom$bf_05_fwe <- .05/nrow(boom)

##### Add  Features 

boom$feature_group <- apply( boom %>% select(feature, feature_group), 1, sort_the_drivers)
boom$feature_group <- ifelse(boom$feature == "hla_lilac_mut_hla", "somatic", boom$feature_group)
boom$plot_est <- ifelse(boom$model %in% c("bor", "relapse", "surv_at_t"), exp(boom$est), 1/exp(boom$est))
boom$big_group <- unlist(lapply( boom$feature_group, big_grouper))
boom$little_group <- unlist(lapply( boom$feature_group, little_grouper))
boom$Direction <- ifelse(boom$plot_est > 1, "Better", "Worse")
boom$Group <- apply( boom %>% select(feature, feature_group), 1, grouper)
boom <- boom %>% filter( plot_est < 5, plot_est > .2)

grouper <- function( big_group, cor_tcell, cor_prolif, cor_tgfb ){
    if( big_group != "RNA"){ big_group } 
    else {
      if( cor_tcell > .5){  "RNA: T-cell"} 
      else if (cor_tgfb > .5){  "RNA: TGFB"} 
      else if (cor_prolif > .5){ "RNA: Proliferation" } 
      else { "RNA: Remaining" }
    }
}

boom <-
boom %>% 
  rowwise() %>% 
  mutate(discovery_group = grouper( big_group, cor_tcell, cor_prolif, cor_tgfb )) %>% 
  mutate(discovery_group = factor(discovery_group, 
         levels = c("Somatic", "RNA: T-cell", "RNA: TGFB", "RNA: Proliferation", "RNA: Remaining",
                    "CNV/SVs", "Clinical", "HLA" ))) %>% 
  ungroup() %>% 
  mutate(Type = discovery_group) %>% 
  mutate(big_group = ifelse(big_group %in% c("CNV", "SVs") , "CNV/SVs", big_group)) %>% 
  mutate(big_group = factor(big_group, levels = c('Somatic','RNA','CNV/SVs','Clinical','HLA'))) %>% 
  mutate(pval_significance = ifelse(p_val < by_05_fdr, "sign", "non-sign"))

### 2 - Save for more analysis and figures

saveRDS( boom, paste0(TMP_DIR,'exhaustive-plots-base.Rds'))

### 3 - Save subset of data as supplementary table

supplement_table <- 
boom %>% 
  select(dataset, model, model_type, Type, feature, est, se, p_val, covariates, contains("cor_"), contains("05")) %>% 
  mutate(feature = gsub("mariathan_", "mariathasan_", feature)) %>% 
  mutate( covariates = ifelse( covariates == "residuals", "age_biopsy_purity_tissue_tmb_tcell_pretreat", as.character(covariates))) %>% 
  mutate( covariates = ifelse( covariates == "residuals2", "age_biopsy_purity_tissue_tmb_tcell_pretreat_tgfb", as.character(covariates)))

fwrite(supplement_table, paste0(FIG_FINAL_DIR, "Table_S3.csv"))

paste0(FIG_FINAL_DIR, "Table_S3.csv")
